View		Reactivate zooming.
svn path=/trunk/koffice/; revision=654249

+++ kspread/View.h
@@ -316,7 +316,7 @@ public Q_SLOTS:
     void togglePageBorders( bool );
     void toggleProtectSheet( bool );
     void toggleProtectDoc( bool );
-    void viewZoom( KoZoomMode::Mode mode, int zoom );
+    void viewZoom( KoZoomMode::Mode mode, double zoom );
     void find();
     void findNext();
     void findPrevious();
